+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [#1032](https://github.com/Azure/azure-storage-fuse/issues/1032) `mount all` CLI params parsing fix when no `config-file` and `tmp-path` is provided.
 - [#1015](https://github.com/Azure/azure-storage-fuse/issues/1015) Default value of `ignore-open-flags` config parameter changed to `true`.
 - [#1038](https://github.com/Azure/azure-storage-fuse/issues/1038) Changing default daemon permissions.
+- [#1036](https://github.com/Azure/azure-storage-fuse/issues/1036) Fix to avoid panic when $HOME dir is not set.
+- [#1036](https://github.com/Azure/azure-storage-fuse/issues/1036) Respect --default-working-dir cli param and use it as default log file path.
 
 ## 2.0.1 (2022-12-02)
 - Copy of GA release of Blobfuse2. This release was necessary to ensure the GA version resolves ahead of the preview versions.