+++ CHANGELOG.md
@@ -9,9 +9,7 @@
 - [#1038](https://github.com/Azure/azure-storage-fuse/issues/1038) Changing default daemon permissions.
 - [#1036](https://github.com/Azure/azure-storage-fuse/issues/1036) Fix to avoid panic when $HOME dir is not set.
 - [#1036](https://github.com/Azure/azure-storage-fuse/issues/1036) Respect --default-working-dir cli param and use it as default log file path.
-
-**Breaking Changes**
-- If version check fails due to public container not reachable, mount/mountall/mountv1 command used to terminate. From this release it will just emit a error log.
+- If version check fails due to network issues, mount/mountall/mountv1 command used to terminate. From this release it will just emit an error log and mount will continue.
 
 
 ## 2.0.1 (2022-12-02)