+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [#1036](https://github.com/Azure/azure-storage-fuse/issues/1036) Fix to avoid panic when $HOME dir is not set.
 - [#1036](https://github.com/Azure/azure-storage-fuse/issues/1036) Respect --default-working-dir cli param and use it as default log file path.
 - If version check fails due to network issues, mount/mountall/mountv1 command used to terminate. From this release it will just emit an error log and mount will continue.
+- If default work directory does not exists, mount shall create it before daemonizing.
 
 
 ## 2.0.1 (2022-12-02)